refactored code to have routes
