+++ index.js
@@ -1,70 +1,6 @@
-const express = require("express");
-const helmet = require("helmet");
-const cors = require("cors");
-const bcrypt = require("bcryptjs");
+const server = require("./api/server.js");
 
-const db = require("./data/dbConfig.js");
-const Users = require("./users/users-model.js");
-const protected = require("./auth/protected-middleware.js");
-
-const server = express();
-
-server.use(helmet());
-server.use(express.json());
-server.use(cors());
-
-server.get("/", (req, res) => {
-  res.send("It's alive!");
-});
-//creates user and changes password to hash
-server.post("/api/register", (req, res) => {
-  let user = req.body;
-
-  const hash = bcrypt.hashSync(user.password, 8);
-
-  user.password = hash;
-
-  Users.add(user)
-    .then(saved => {
-      res.status(201).json(saved);
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-//logs in registered user and returns logged in message
-server.post("/api/login", (req, res) => {
-  let { username, password } = req.body;
-
-  Users.findBy({ username })
-    .first()
-    .then(user => {
-      //
-      if (user && bcrypt.compareSync(password, user.password)) {
-        res.status(200).json({
-          message: `Welcome ${user.username}! You are now logged in! `
-        });
-      } else {
-        res.status(401).json({ message: "YOU SHALL NOT PASS" });
-      }
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-//user must be logged in to view this,
-server.get("/api/users", protected, (req, res) => {
-  Users.find()
-    .then(users => {
-      res.json(users);
-    })
-    .catch(err => res.send(err));
-});
-
-const port = process.env.PORT || 5000;
-
+const port = 5000;
 server.listen(port, () => {
-  console.log(`Server is live on ${port}`);
+  console.log(`\n** Running on port ${port} **\n`);
 });
